requirements.txt
boto3
python-dotenv

# === Assignment 1 - legacy ===
# fastapi 
# uvicorn
# httpx
# aiohttp
